Quiz Application
We are developing a Python Quiz Application with a Tkinter-based frontend and a MySQL backend. The application offers quizzes in four distinct domains, each comprising 30 questions. Players complete the quiz, and upon completion, their final scores, along with other player details, are recorded in the scoreboard table within the backend database.